requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-ollama>=0.2.0
python-dotenv>=1.0.1
chromadb>=1.5.0
flet>=0.27.0
